fix(context-fabric): codex round-2 findings F1-F4 (CHAOS-3781)
F1 HIGH -- devhealthfacts scanned git_pull_requests.number (UInt32) into
an *int64, which clickhouse-go rejects outright, so every live row failed
Scan and the provider returned no pull-request facts at all. CHAOS-3789
fixed exactly this in devhealthsource; the defect survived here because
this package's fixtures modeled the column as int64, so its tests agreed
with the bug.

Extending the parity guard to devhealthfacts immediately found a SECOND
live instance the review had not reported: source_health scanned
items_synced (UInt32) and duration_ms (UInt64) into *int64. Both now
convert in SQL with toInt64(...), matching every other numeric projection
in the package.

The production column snapshot is now declared ONCE, in the new
devhealthschema package, and every guard and fixture in both packages
renders from it -- devhealthsource's guard, its freshness check, its three
org-isolation fixtures, and devhealthfacts' new guard. A fixture cannot
catch a fixture's own error; two artifacts from one declaration cannot
disagree. The declaration carries each table's production ENGINE (readers
use FINAL, which errors against a plain MergeTree) and production column
POSITION order (so a fixture is a positional replica).

F2 MED -- reuse lookup keyed from the CLAMPED request while save keyed
from the wire request, so whenever clamping fired the two disagreed and an
identical request could never reuse its own answer, with the lookup key
drifting with now. Both sides now key pre-clamp. The round-1 test missed
this because its as-of was safely in the past; the new guard puts the
instant inside the skew tolerance, the only place it is observable.

F3 MED -- composed grain counted only SourceAvailable providers, but the
bundle RETAINS facts from truncated and fact-bearing stale ones. A
truncated daily rollup contributed its facts while its grain was dropped,
so a mixed answer labeled instant and overstated its own precision.
Contribution now means facts-retained, via a shared factsRetained
predicate so retention and grain cannot drift apart again.

F4 MED -- the org-isolation fixtures declared their own types
(DateTime64(6,'UTC') where production is mostly DateTime64(3), repo_id as
String where production is UUID), so they proved isolation against a
schema production does not have. They now render from the shared
declaration.

Every finding has a red-green guard, each verified by reverting the fix
and confirming the guard fails for the RIGHT reason -- the first attempt
at F1's red failed on a DDL error instead of the scan, which is why the
message was checked rather than the red taken at face value.

+++ internal/contextfabric/devhealthfacts/pullrequests_test.go
@@ -23,7 +23,9 @@ func reviewSubject(reviewID string) contextfabric.SubjectRef {
 func TestPullRequestsProviderHappyPath(t *testing.T) {
 	t.Parallel()
 	client := &fakeClient{tables: []fakeTable{
-		{match: "FROM git_pull_requests", rows: [][]any{{"repo-1", int64(1042), "open"}}},
+		// uint32, matching the production column type -- an int64 fixture here
+		// is what let the reader's own int64 scan pass for so long.
+		{match: "FROM git_pull_requests", rows: [][]any{{"repo-1", uint32(1042), "open"}}},
 	}}
 	provider := findProvider(t, devhealthfacts.NewProviders(client), contextfabric.FactPullRequests)
 	result, err := provider.ReadFacts(context.Background(), storage.Principal{OrgID: "org-1"}, contextfabric.FactQuery{
